Add endpoint to fetch project assignment candidates
Introduced a new API endpoint in ProjectAssignmentsController to retrieve user profiles assigned to a specific project assignment. The logic for fetching candidates was moved from UserProfileService to ProjectAssignmentService, and the corresponding interface methods were updated.
