+++ src/SkillBridge/Services/ProjectAssignment/ProjectAssignmentService.cs
@@ -493,4 +493,11 @@ public class ProjectAssignmentService : IProjectAssignmentService
         return results.Select(p => _mapper.Map<ProjectAssignmentResponse>(p));
 
     }
+    public async Task<List<Models.Entities.UserProfile>> GetByAssigmentId(Guid projectId)
+    {
+        var candidates = await _dbContext.UserProjectAssignments.Include(x => x.UserProfile).Where(upa => upa.ProjectAssignmentId == projectId).Select(x => x.UserProfile)
+            .ToListAsync();
+
+        return candidates;
+    }
 }
